+++ packages/services/f-braze-adapter/src/services/utils/_tests/transformCardData.test.js
@@ -24,19 +24,31 @@ const extractedCardId = '__EXTRACTED_CARD_ID__';
 const description = '__DESCRIPTION__';
 const subtitle = '__SUBTITLE__';
 const target = { attribute: '_self' };
+const discountPercentage = '__DISCOUNT_PERCENTAGE__';
+const earnedStamps = '__EARNED_STAMPS__';
+const expiryDate = '__EXPIRY_DATE__';
+const expiryLine = '__EXPIRY_LINE__';
+const isReadyToClaim = '__IS_READY_TO_CLAIM__';
+const totalRequiredStamps = '__TOTAL_REQUIRED_STAMPS__';
 
 getCardUrlTarget.mockReturnValue(target);
 extractCardId.mockReturnValue(extractedCardId);
 
 const extras = {
     button_1: button, // eslint-disable-line camelcase
+    discount_percentage: discountPercentage, // eslint-disable-line camelcase
+    earned_stamps: earnedStamps, // eslint-disable-line camelcase
+    expiry_date: expiryDate, // eslint-disable-line camelcase
+    expiry_line: expiryLine, // eslint-disable-line camelcase
     icon_1: icon, // eslint-disable-line camelcase
     image_1: image, // eslint-disable-line camelcase
+    is_ready_to_claim: isReadyToClaim, // eslint-disable-line camelcase
     line_1: line1, // eslint-disable-line camelcase
     line_2: line2, // eslint-disable-line camelcase
     order,
     custom_card_type: type, // eslint-disable-line camelcase
     voucher_code: voucherCode, // eslint-disable-line camelcase
+    total_required_stamps: totalRequiredStamps, // eslint-disable-line camelcase
     updated
 };
 
@@ -53,13 +65,19 @@ const rawCard = {
 
 const reformattedCard = {
     description: [line1, line2],
+    discountPercentage,
+    earnedStamps,
+    expiryDate,
+    expiryLine,
     id,
     extractedCardId,
     url,
     icon,
     ctaText: button,
     image,
+    isReadyToClaim,
     title,
+    totalRequiredStamps,
     subtitle: description,
     order,
     voucherCode,